Added ability to find 10 "nearest" locations, based on distance from a specified point (WIP). Also some slight improvements such as moving api endpoint to appsettings.

+++ LeedsBeerQuest/LeedsBeerQuest.Api/Models/BeerEstablishment.cs
@@ -28,7 +28,7 @@ namespace LeedsBeerQuest.Api.Models
 
     public struct Location
     {
-        public decimal Lat { get; set; }
-        public decimal Long { get; set; }
+        public double Lat { get; set; }
+        public double Long { get; set; }
     }
 }
